+++ app/src/main/java/ahgpoug/com/specshelper/adapters/MotherboardRecyclerAdapter.java
@@ -2,6 +2,7 @@ package ahgpoug.com.specshelper.adapters;
 
 
 import android.content.Context;
+import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -11,8 +12,13 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.afollestad.materialdialogs.MaterialDialog;
+
+import java.io.Serializable;
 import java.util.ArrayList;
 
+import ahgpoug.com.specshelper.ActivityAddNewCpu;
+import ahgpoug.com.specshelper.ActivityAddNewMotherboard;
 import ahgpoug.com.specshelper.util.DataBaseHelper;
 import ahgpoug.com.specshelper.util.Globals;
 import ahgpoug.com.specshelper.objects.Motherboard;
@@ -136,19 +142,39 @@ public class MotherboardRecyclerAdapter extends RecyclerView.Adapter<Motherboard
             holder.name.getRootView().setOnLongClickListener(new View.OnLongClickListener() {
                 @Override
                 public boolean onLongClick(View view) {
-                    String query = "DELETE FROM motherboard WHERE id = " + values.get(position).getId();
+                    new MaterialDialog.Builder(context)
+                            .title("Действия")
+                            .items(new String[] {"Изменить", "Удалить"})
+                            .itemsCallback(new MaterialDialog.ListCallback() {
+                                @Override
+                                public void onSelection(MaterialDialog dialog, View view, int which, CharSequence text) {
+                                    if (text.equals("Удалить")){
+                                        String query = "DELETE FROM motherboard WHERE id = " + values.get(position).getId();
 
-                    DataBaseHelper mDatabaseHelper = new DataBaseHelper(context);
-                    SQLiteDatabase db = mDatabaseHelper.getWritableDatabase();
+                                        DataBaseHelper mDatabaseHelper = new DataBaseHelper(context);
+                                        SQLiteDatabase db = mDatabaseHelper.getWritableDatabase();
 
-                    db.execSQL(query);
+                                        db.execSQL(query);
 
-                    if ((Globals.cart.getMbID() == values.get(position).getId()))
-                        Globals.cart.setMbID(-1);
+                                        if ((Globals.cart.getMbID() == values.get(position).getId()))
+                                            Globals.cart.setMbID(-1);
 
-                    values.remove(position);
-                    notifyDataSetChanged();
-                    Toast.makeText(context, "Удалено", Toast.LENGTH_LONG).show();
+                                        values.remove(position);
+                                        notifyDataSetChanged();
+                                        Toast.makeText(context, "Удалено", Toast.LENGTH_LONG).show();
+                                    } else {
+                                        Intent intent = new Intent(context, ActivityAddNewMotherboard.class);
+                                        intent.putExtra("mb", (Serializable) values.get(position));
+                                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                        context.startActivity(intent);
+
+                                        if ((Globals.cart.getMbID() == values.get(position).getId()))
+                                            Globals.cart.setMbID(-1);
+                                        notifyDataSetChanged();
+                                    }
+                                }
+                            })
+                            .show();
                     return false;
                 }
             });
